+++ src/Mapbender/CoreBundle/Resources/public/mapbender.element.coordinatesdisplay.js
@@ -85,7 +85,7 @@
                     ;
                 };
             } else {
-                controlOptions.element = $('#coordinatesdisplay', this.element).get(0);
+                controlOptions.element = $('.display-area', this.element).get(0);
             }
             if (this.control) {
                 this.mbMap.map.olMap.removeControl(this.control);