+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+* **v3.0.8-beta3**
+  - [Beta regression] Restore Wms Loader function
+  - [Beta regression] Fix error on SimpleSearch centering
+  - [Beta regression] Fix behavior of LayerTree opacity menu when reducing opacity to zero
+  - [Beta regression] Fix PrintClient selection broken in Firefox
+  - [Beta regression] Fix ZoomBar missing highlight on activated zoom box icon
+  - Update `mbPrintClient.printDigitizerFeature` API ([PR#1123](https://github.com/mapbender/mapbender/pull/1123), [Digitizer PR#69](https://github.com/mapbender/mapbender-digitizer/pull/69))
+  - Fix broken display of Buttons with 'Coordinates (FontAwesome)' icon assigned
+  - Fix missing gap in ScaleDisplay
+  - Fix PrintClient errors when switching template while selection rectangle is disabled (sidepane / element mode)
+
 * **v3.0.8-beta2**
   - [Beta regression] Fix button group misalginment in Application "Layersets" tab
   - [Beta regression] Fix layout breakage in "Vendorspecifics" area of WMS instance backend